Fix the internal PMIx installation - matching changes have been upstreamed

+++ opal/mca/pmix/pmix2x/pmix/src/include/pmix_globals.h
@@ -68,7 +68,7 @@ PMIX_CLASS_DECLARATION(pmix_namelist_t);
 
 /* define a command type for communicating to the
  * pmix server */
-typedef uint32_t pmix_cmd_t;
+typedef uint8_t pmix_cmd_t;
 
 /* define some commands */
 #define PMIX_REQ_CMD             0
